+++ src/components/Timer/index.ts
@@ -1,0 +1,3 @@
+import Timer from "./Timer";
+
+export { Timer };
