--- src/components/Timer/timer.tsx
@@ -1,87 +1,0 @@
-import React, { useEffect, useState } from "react";
-import { TimerDisplay, DescriptionField, ProjectField } from "../index";
-import useSession from "./useSession";
-import { formatSeconds, getUnixTime } from "../../utils";
-import useClock from "../../hooks/useClock/useClock";
-import { format } from "date-fns";
-import { Button, ButtonGroup, TextField } from "@material-ui/core";
-
-const timeNow = () => {
-  return format(Date.now(), "s.SSS");
-};
-
-const Timer = () => {
-  const [time, setTime] = useState<number>(15 * 60);
-  const [isInProgress, setIsInProgress] = useState<boolean>(false);
-  const [buttonText, setButtonText] = useState<string>("Start");
-  const { start, stop, ticks } = useClock();
-  const { startSession, endSession } = useSession();
-  const [description, setDescription] = useState("");
-  const [project, setProject] = useState("");
-
-  // Main clock effect
-  useEffect(() => {
-    if (isInProgress) {
-      setTime((prev) => prev - 1);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [ticks]);
-
-  // Document Title updates
-  useEffect(() => {
-    if (isInProgress) {
-      document.title = formatSeconds(time).join(":");
-    } else {
-      document.title = "Pomi";
-    }
-  }, [isInProgress, time]);
-
-  const onStartStopClick = () => {
-    if (!isInProgress) {
-      // Start Timer
-      //setTime(15 * 60);
-      start();
-      setIsInProgress(true);
-      setButtonText("Stop");
-      startSession(Date.now(), project, description);
-    } else {
-      setIsInProgress(false);
-      setButtonText("Start");
-      stop();
-      setTime(60 * 15);
-      // TODO: This needs to get refactored and go elsewhere
-      const endTime = Date.now();
-      const newSession = endSession(endTime);
-      const newId = getUnixTime(newSession.date, newSession.startTime);
-      const savedSessions = JSON.parse(
-        window.localStorage.getItem("sessions") || "{}"
-      );
-      savedSessions[newId] = newSession;
-      window.localStorage.setItem("sessions", JSON.stringify(savedSessions));
-      alert(JSON.stringify({ [newId]: newSession }, null, 5));
-    }
-  };
-
-  return (
-    <div id="timer">
-      <ProjectField
-        value={project}
-        onChange={(e) => setProject(e.target.value)}
-        disabled={isInProgress}
-      />
-      <DescriptionField
-        value={description}
-        onChange={(e) => setDescription(e.target.value)}
-        disabled={isInProgress}
-      />
-      <TimerDisplay time={time} setTime={setTime} isInProgress={isInProgress} />
-      <ButtonGroup fullWidth orientation="vertical">
-        <Button variant="contained" color="primary" onClick={onStartStopClick}>
-          {buttonText}
-        </Button>
-      </ButtonGroup>
-    </div>
-  );
-};
-
-export default Timer;
