+++ src/components/index.ts
@@ -8,6 +8,7 @@ import ProjectSearch from "./Inputs/Input";
 import Card from "./Card";
 import Header from "./Header";
 import { DescriptionField, ProjectField } from "./Inputs";
+import { Timer } from "./Timer";
 
 export {
   /* PLOP_INJECT_EXPORT */
@@ -22,4 +23,5 @@ export {
   Header,
   DescriptionField,
   ProjectField,
+  Timer,
 };
