+++ src/hooks/index.ts
@@ -1,0 +1,4 @@
+import useClock from "./useClock/useClock";
+import useSession from "./useSession/useSession";
+
+export { useClock, useSession };
